Added RSS feeds for blog posts.
From now all the pages inside the blog/ dir are indexed in an RSS feed
located at /blog.feed and blog.it.feed for the italian version.
All the blog posts must have both created_at and modified_at meta
information because the first is needed to create the permalink and the
second to sort the posts in the feeds.
Updated default.template to correctly include the feed link in all
pages.