Added date and times localization support to I18nTag.
The I18nTag class now provides a 'l' tag which can be used to localize a date or time based on a default or a specified rule.
Added necessary rules in the locales/*.yml files.
Updated blog/default.template to use that capability to display publication date.
Updated CSSes to better display the publication date.